Add specs to test #my_count and #my_map methods
This commit adds specs to test #my_count method when there is no block
given, when there is an argument given as a literal or a block.

This commit also test #my_map method when there is no block given, when there
is a block given, when there is an argument given as a block and in combination
with a range when a block is given.